Cleanup

+++ common/SaveHelper.java
@@ -44,9 +44,8 @@ public class SaveHelper {
 			File file2 = new File(this.playersDirectory, UniqueID + ".dat");
 			File file3 = new File(this.playersDirectory, UniqueID + ".dat.backup");
 			CompressedStreamTools.writeCompressed(nbttagcompound, new FileOutputStream(file1));
-			if (file2.exists()) {
-				file2.renameTo(file3);
-			}
+			if (file3.exists()) file3.delete();
+			if (file2.exists()) file2.renameTo(file3);			
 			file1.renameTo(file2);
 		} catch (Exception exception) {
 			exception.printStackTrace();
